プレイヤー関連の処理をStopMgrに対応させました。

+++ Game/Player/Player.h
@@ -39,13 +39,13 @@ private:
 	const float MOVE_SPEED_CREEPING = 0.15f;
 
 	//発射の遅延
-	int m_shotDelay;
-	const int SHOT_DELAY = 20;
+	float m_shotDelay;
+	const float SHOT_DELAY = 20;
 
 	//心音のタイマー
-	int m_heatbeatTimer;
-	const int HEARTBEAT_TIMER = 45;
-	const int HEARTBEAT_TIMER_FOUND = 32;
+	float m_heatbeatTimer;
+	const float HEARTBEAT_TIMER = 45;
+	const float HEARTBEAT_TIMER_FOUND = 32;
 	SoundData m_heartbeatSE;
 
 	bool m_isFoundToEnemy;
